+++ services/121-service/module-dependencies.md
@@ -119,6 +119,18 @@ graph LR
   TransactionJobProcessorsModule-->FinancialServiceProvidersModule
   TransactionJobProcessorsModule-->EventsModule
   TransactionJobProcessorsModule-->MessageTemplateModule
-  FinancialServiceProviderCallbackJobProcessorsModule-->RedisModule
-  FinancialServiceProviderCallbackJobProcessorsModule-->SafaricomModule
+  CommercialBankEthiopiaReconciliationModule-->CommercialBankEthiopiaModule
+  CommercialBankEthiopiaReconciliationModule-->RegistrationsModule
+  ExcelReconcilicationModule-->ActionsModule
+  ExcelReconcilicationModule-->TransactionsModule
+  ExcelReconcilicationModule-->ExcelModule
+  ExcelReconcilicationModule-->RegistrationsModule
+  IntersolveVisaReconciliationModule-->IntersolveVisaModule
+  IntersolveVoucherReconciliationModule-->IntersolveVoucherModule
+  IntersolveVoucherReconciliationModule-->ProgramModule
+  IntersolveVoucherReconciliationModule-->ProgramFinancialServiceProviderConfigurationsModule
+  SafaricomReconciliationModule-->SafaricomModule
+  SafaricomReconciliationModule-->RedisModule
+  SafaricomReconciliationModule-->TransactionsModule
+  SafaricomReconciliationModule-->QueuesRegistryModule
 ```
