+++ services/121-service/src/payments/fsp-integration/excel/excel.service.ts
@@ -11,19 +11,14 @@ import {
   FspTransactionResultDto,
   PaTransactionResultDto,
 } from '@121-service/src/payments/dto/payment-transaction-result.dto';
-import { ReconciliationFeedbackDto } from '@121-service/src/payments/dto/reconciliation-feedback.dto';
 import { TransactionRelationDetailsDto } from '@121-service/src/payments/dto/transaction-relation-details.dto';
 import { ExcelFspInstructions } from '@121-service/src/payments/fsp-integration/excel/dto/excel-fsp-instructions.dto';
 import { FinancialServiceProviderIntegrationInterface } from '@121-service/src/payments/fsp-integration/fsp-integration.interface';
-import { ReconciliationReturnType } from '@121-service/src/payments/interfaces/reconciliation-return-type.interface';
 import { TransactionReturnDto } from '@121-service/src/payments/transactions/dto/get-transaction.dto';
 import { TransactionStatusEnum } from '@121-service/src/payments/transactions/enums/transaction-status.enum';
 import { TransactionsService } from '@121-service/src/payments/transactions/transactions.service';
-import { ProgramFinancialServiceProviderConfigurationEntity } from '@121-service/src/program-financial-service-provider-configurations/entities/program-financial-service-provider-configuration.entity';
 import { ProgramFinancialServiceProviderConfigurationRepository } from '@121-service/src/program-financial-service-provider-configurations/program-financial-service-provider-configurations.repository';
 import { ProgramEntity } from '@121-service/src/programs/program.entity';
-import { ImportStatus } from '@121-service/src/registration/dto/bulk-import.dto';
-import { MappedPaginatedRegistrationDto } from '@121-service/src/registration/dto/mapped-paginated-registration.dto';
 import { RegistrationsPaginationService } from '@121-service/src/registration/services/registrations-pagination.service';
 import { FileImportService } from '@121-service/src/utils/file-import/file-import.service';
 
@@ -169,7 +164,7 @@ export class ExcelService
     );
   }
 
-  private joinRegistrationsAndTransactions(
+  public joinRegistrationsAndTransactions(
     orderedRegistrations: Awaited<
       ReturnType<RegistrationsPaginationService['getRegistrationsChunked']>
     >,
@@ -248,227 +243,5 @@ export class ExcelService
       );
     }
     return matchColumn;
-  }
-
-  public async processReconciliationData({
-    file,
-    payment,
-    programId,
-    fspConfigs,
-  }: {
-    file: Express.Multer.File;
-    payment: number;
-    programId: number;
-    fspConfigs: ProgramFinancialServiceProviderConfigurationEntity[];
-  }): Promise<ReconciliationReturnType[]> {
-    const maxRecords = 10000;
-    const validatedExcelImport = await this.fileImportService.validateCsv(
-      file,
-      maxRecords,
-    );
-
-    // First set up unfilled feedback object based on import rows ..
-    const crossFspConfigImportResults: ReconciliationReturnType[] = [];
-    for (const row of validatedExcelImport) {
-      const resultRow = new ReconciliationReturnType();
-      resultRow.feedback = new ReconciliationFeedbackDto();
-      resultRow.feedback = {
-        ...row,
-        importStatus: ImportStatus.notFound,
-        referenceId: null,
-        message: null,
-      };
-      resultRow.programFinancialServiceProviderConfigurationId = undefined;
-      resultRow.transaction = undefined;
-      crossFspConfigImportResults.push(resultRow);
-    }
-
-    // .. then loop over fspConfigs to update rows where matched
-    for await (const fspConfig of fspConfigs) {
-      const matchColumn = await this.getImportMatchColumn(fspConfig.id);
-      const importResultForFspConfig = await this.reconciliatePayments({
-        programId,
-        payment,
-        validatedExcelImport,
-        fspConfig,
-        matchColumn,
-      });
-      // Convert the array into a map for increased performace (hashmap lookup)
-      const importResultForFspConfigMap = new Map(
-        importResultForFspConfig.map((item) => [
-          item.feedback[matchColumn],
-          item,
-        ]),
-      );
-
-      // .. then loop over each row of the original import to update if a match has been found with this fspConfig
-      crossFspConfigImportResults.forEach((row, index) => {
-        const importResultForFspConfigRow = importResultForFspConfigMap.get(
-          row.feedback[matchColumn],
-        );
-        if (
-          importResultForFspConfigRow?.feedback.importStatus !==
-          ImportStatus.notFound
-        ) {
-          crossFspConfigImportResults[index] = importResultForFspConfigRow!;
-        }
-      });
-    }
-    return crossFspConfigImportResults;
-  }
-
-  private async reconciliatePayments({
-    programId,
-    payment,
-    validatedExcelImport,
-    fspConfig,
-    matchColumn,
-  }: {
-    programId: number;
-    payment: number;
-    validatedExcelImport: object[];
-    fspConfig: ProgramFinancialServiceProviderConfigurationEntity;
-    matchColumn: string;
-  }): Promise<ReconciliationReturnType[]> {
-    const registrationsForReconciliation =
-      await this.getRegistrationsForReconciliation(
-        programId,
-        payment,
-        matchColumn,
-        fspConfig.id,
-      );
-    if (!registrationsForReconciliation?.length) {
-      return [];
-    }
-    const lastTransactions = await this.transactionsService.getLastTransactions(
-      programId,
-      payment,
-      undefined,
-      undefined,
-      fspConfig.id,
-    );
-    // Join registration data with the imported CSV records
-    return this.joinRegistrationsAndImportRecords(
-      registrationsForReconciliation,
-      validatedExcelImport,
-      matchColumn,
-      lastTransactions,
-      fspConfig.id,
-    );
-  }
-
-  private async getRegistrationsForReconciliation(
-    programId: number,
-    payment: number,
-    matchColumn: string,
-    programFinancialServiceProviderConfigurationId: number,
-  ): Promise<MappedPaginatedRegistrationDto[]> {
-    const qb =
-      this.registrationsPaginationService.getQueryBuilderForFspInstructions({
-        programId,
-        payment,
-        programFinancialServiceProviderConfigurationId,
-        financialServiceProviderName: FinancialServiceProviders.excel,
-      });
-    // log query
-    const chunkSize = 400000;
-    return await this.registrationsPaginationService.getRegistrationsChunked(
-      programId,
-      {
-        select: [matchColumn, 'referenceId', 'id'],
-        path: '',
-      },
-      chunkSize,
-      qb,
-    );
-  }
-
-  private joinRegistrationsAndImportRecords(
-    registrations: Awaited<
-      ReturnType<ExcelService['getRegistrationsForReconciliation']>
-    >,
-    importRecords: object[],
-    matchColumn: string,
-    existingTransactions: TransactionReturnDto[],
-    fspConfigId: number,
-  ): ReconciliationReturnType[] {
-    // First order registrations by referenceId to join amount from transactions
-    const registrationsOrderedByReferenceId = registrations.sort((a, b) =>
-      a.referenceId.localeCompare(b.referenceId),
-    );
-    const registrationsWithAmount = this.joinRegistrationsAndTransactions(
-      registrationsOrderedByReferenceId,
-      existingTransactions,
-      ['id', 'referenceId', matchColumn],
-    );
-
-    // Then order registrations and importRecords by matchColumn to join them
-    const importRecordsOrdered = importRecords.sort((a, b) =>
-      a[matchColumn]?.localeCompare(b[matchColumn]),
-    );
-    const registrationsOrdered = registrationsWithAmount.sort((a, b) =>
-      (a[matchColumn] as string).localeCompare(b[matchColumn] as string),
-    );
-
-    const resultFeedback: ReconciliationReturnType[] = [];
-    for (const record of importRecordsOrdered) {
-      let transaction: PaTransactionResultDto | null = null;
-      let importStatus = ImportStatus.notFound;
-
-      if (
-        ![TransactionStatusEnum.success, TransactionStatusEnum.error].includes(
-          record[this.statusColumnName]?.toLowerCase(),
-        )
-      ) {
-        const errors = `The 'status' column is either missing or contains unexpected values. It should only contain 'success' or 'error'.`;
-        throw new HttpException({ errors }, HttpStatus.NOT_FOUND);
-      } else if (record[matchColumn] === undefined) {
-        const errors = `The match column '${matchColumn}' is missing.`;
-        throw new HttpException({ errors }, HttpStatus.NOT_FOUND);
-      }
-
-      // find registration with matching matchColumn value
-      const matchedRegistration = registrationsOrdered.find(
-        (r) => r[matchColumn] === record[matchColumn],
-      );
-      if (matchedRegistration) {
-        transaction = this.createTransactionResult(matchedRegistration, record);
-        importStatus =
-          transaction.status === TransactionStatusEnum.success
-            ? ImportStatus.paymentSuccess
-            : ImportStatus.paymentFailed;
-      }
-      resultFeedback.push({
-        feedback: {
-          referenceId: (matchedRegistration?.referenceId as string) ?? null,
-          status: transaction?.status ?? null,
-          message: transaction?.message ?? null,
-          importStatus,
-          [matchColumn]: record[matchColumn],
-        },
-        programFinancialServiceProviderConfigurationId: matchedRegistration
-          ? fspConfigId
-          : undefined,
-        transaction: transaction || undefined,
-      });
-    }
-
-    return resultFeedback;
-  }
-
-  public createTransactionResult(
-    registrationWithAmount: ExcelFspInstructions,
-    importResponseRecord: any,
-  ): PaTransactionResultDto {
-    return {
-      referenceId: registrationWithAmount.referenceId,
-      registrationId: registrationWithAmount.id,
-      fspName: FinancialServiceProviders.excel,
-      status: importResponseRecord[
-        this.statusColumnName
-      ]?.toLowerCase() as TransactionStatusEnum,
-      calculatedAmount: registrationWithAmount.amount,
-      message: null,
-    };
   }
 }
