+++ services/121-service/src/payments/transactions/transactions.service.ts
@@ -302,7 +302,7 @@ export class TransactionsService {
     transactionResults: PaTransactionResultDto[],
     transactionRelationDetails: TransactionRelationDetailsDto,
   ): Promise<void> {
-    // NOTE: this method is currently only used for the import-fsp-reconciliation use case and assumes:
+    // NOTE: this method is currently only used for the import-excel-reconciliation use case and assumes:
     // 1: only 1 program financial service provider id
     // 2: no notifications to send
     // 3: no payment count to update (as it is reconciliation of existing payment)
