+++ services/121-service/test/helpers/program.helper.ts
@@ -214,7 +214,7 @@ export async function importFspReconciliationData(
   const csvString = jsonArrayToCsv(reconciliationData);
   const buffer = Buffer.from(csvString, 'utf-8');
   return await getServer()
-    .post(`/programs/${programId}/payments/${paymentNr}/fsp-reconciliation`)
+    .post(`/programs/${programId}/payments/${paymentNr}/excel-reconciliation`)
     .set('Cookie', [accessToken])
     .field('Content-Type', 'multipart/form-data')
     .attach('file', buffer, 'reconciliation.csv');
@@ -418,8 +418,8 @@ export async function startCbeValidationProcess(
   accessToken: string,
 ): Promise<request.Response> {
   return await getServer()
-    .post(
-      `/programs/${programId}/financial-service-providers/commercial-bank-ethiopia/account-enquiries/validation`,
+    .put(
+      `/programs/${programId}/financial-service-providers/commercial-bank-ethiopia/account-enquiries`,
     )
     .set('Cookie', [accessToken]);
 }
